+++ pyjsonrpc/http.py
@@ -3,9 +3,15 @@
 
 import urllib2
 import base64
+import BaseHTTPServer
+import SocketServer
+import httplib
+import urllib
+import urlparse
 import rpcrequest
 import rpcresponse
 import rpcerror
+import rpclib
 from rpcjson import json
 
 
@@ -120,4 +126,90 @@ class HttpClient(object):
 
         return self._Method(http_client_instance = self, method = method)
 
+
+class ThreadingHttpServer(SocketServer.ThreadingMixIn, BaseHTTPServer.HTTPServer):
+    """
+    Threading HTTP Server
+    """
+    pass
+
+
+class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
+    """
+    HttpRequestHandler for JSON-RPC-Requests
+
+    Info: http://www.simple-is-better.org/json-rpc/transport_http.html
+    """
+
+    protocol_version = "HTTP/1.1"
+    #server_version = "BaseHTTP/" + __version__
+
+
+    def set_content_type_json(self):
+        """
+        Setzt den Content-Type-Header auf "application/json"
+        """
+
+        self.send_header("Content-Type", "application/json")
+
+
+    def set_content_length(self, length):
+        """
+        Setzt den Content-Lenght-Header
+        """
+
+        self.send_header("Content-Length", str(length))
+
+
+    def do_GET(self):
+        """
+        Handles HTTP-GET-Request
+        """
+
+        # Parse URL query
+        query = urlparse.parse_qs(urllib.splitquery(self.path)[1])
+
+        # jsonrpc
+        jsonrpc = query.get("jsonrpc")
+        if jsonrpc:
+            jsonrpc = jsonrpc[0]
+
+        # id
+        id = query.get("id")
+        if id:
+            id = id[0]
+
+        # method
+        method = query.get("method")
+        if method:
+            method = method[0]
+
+        # params
+        args = []
+        kwargs = {}
+        params = query.get("params")
+        if params:
+            params = json.loads(params[0])
+            if isinstance(params, list):
+                args = params
+                kwargs = {}
+            elif isinstance(params, dict):
+                args = []
+                kwargs = params
+
+        # Create JSON reqeust string
+        request_dict = rpcrequest.create_request_dict(method, *args, **kwargs)
+        request_dict["jsonrpc"] = jsonrpc
+        request_dict["id"] = id
+        request_json = json.dumps(request_dict)
+
+        # Call
+        response_json = self.call(request_json)
+
+        # Return result
+        self.send_response(code = httplib.OK)
+        self.set_content_type_json()
+        self.set_content_length(len(response_json))
+        self.end_headers()
+        self.wfile.write(response_json)